feat: Sort entrance totem messages by where they appear in the raid.
Problem:
The messages correspond to the order in which the user taps the icon, and not where they appear in the raid.

Solution:
Pull the descriptions into a sorted array and provide sort keys to sort and re-map the description so that the messages are displayed to the user in order to where they appear in the raid.

Demo:
Using Kill (8), Pyramid (0), Worship (4)
![](https://i.polds.dev/lyZcT.png)